# Automate cache handling