### Summary

This PR improves the cache handling behavior and simplifies usage by automating cache management. https://github.com/n3xem/gh-otui/issues/8

### Changes

1. **Automatic cache creation** 
   If no cache is found, it is now automatically generated on the first run.

2. **Background cache refresh**
   If the existing cache is older than 1 hour, a background update is triggered automatically without blocking the main execution.

3. **Removed `--cache` flag**  
   Since cache handling is now automatic, the `--cache` flag has been removed for simplicity.

Additionaly this PR improves performance via parallel fetching.

```sh
# Before
hyperfine -w 3 'gh otui --cache'
Benchmark 1: gh otui --cache
  Time (mean ± σ):      7.746 s ±  0.548 s    [User: 0.242 s, System: 0.084 s]
  Range (min … max):    6.944 s …  8.806 s    10 runs

# After
hyperfine -w 3 'gh otui'      
Benchmark 1: gh otui --cache
  Time (mean ± σ):      5.135 s ±  0.515 s    [User: 0.247 s, System: 0.086 s]
  Range (min … max):    4.593 s …  6.027 s    10 runs
```